pyproject.toml
[tool.poetry]
name = "be"
version = "0.1.0"
description = ""
authors = ["kuca4hk <jakub.kucera.01@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "3.2.19"
djangorestframework = "3.11.2"
python-decouple = "^3.8"
psycopg2-binary = "^2.9.9"
dj-database-url = "^2.1.0"
django-cors-headers = "^4.3.1"
drf-yasg2 = "1.19.4"
django-debug-toolbar = "^4.2.0"
flake8 = "^6.1.0"
pre-commit = "^3.6.0"
pillow = "^10.1.0"
ipython = "^8.20.0"
django-extensions = "3.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
